lint: Fix document formatting so we can build gh pages for them
draft-dnsop-deleg-transport.md needed to be set to [`stand_alone: yes`](https://github.com/cabo/kramdown-rfc/#the-yaml-header) to be able to
run `make gh-pages`.

Abbreviation/title of dnssec and transport documents were updated to be able to differentiate them.

Finally, applying the fix from `make fix-lint` in order to be able to commit the changes.